feat : MarketModel 업데이트(price Type 변경 : Int -> Long )

+++ app/src/main/java/com/leebaeng/ggyulmarket/home/MarketModel.kt
@@ -4,7 +4,7 @@ data class MarketModel(
     val sellerId: String = "",
     val title: String = "",
     val createdAt: Long = 0L,
-    val price: Int = 0,
+    val price: Long = 0L,
     val description: String = "",
     val imgUrl: List<String>? = null,
     val talkCnt: Int? = null,
